tried to make a model loader so i can load blockless models

+++ src/main/java/com/magneticraft2/common/blockentity/general/BaseBlockEntityMagneticraft2.java
@@ -37,6 +37,7 @@ import javax.annotation.Nonnull;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
 /**
  * @author JumpWatch on 01-07-2023
@@ -116,66 +117,53 @@ public abstract class BaseBlockEntityMagneticraft2 extends BlockEntity implement
     protected void setMultiblockController(MultiblockController multiblockController) {
         this.multiblockController = multiblockController;
     }
-    protected void setMultiblock(Multiblock multiblock){
+    protected void setMultiblock(Multiblock multiblock) {
         this.multiblock = multiblock;
     }
-
     protected MultiblockController getMultiblockController() {
         return this.multiblockController;
     }
+
     public void onRightClick() {
         if (this.multiblockController != null) {
-            this.multiblockController.createStructure(level,getBlockPos());
-        }else{
+            this.multiblockController.createStructure(level, getBlockPos());
+        } else {
             createMultiblockController();
         }
     }
-    public void onDestry(){
-        this.multiblockController.destroyStructure(level);
+    public void onDestroy(Level level) {
+        if (this.multiblockController != null) {
+            LOGGER.info("onDestroy triggered");
+            this.multiblockController.destroyStructure(level);
+        }
     }
     public boolean matchesStructure(Level world, BlockPos pos, @NotNull MultiblockStructure structure, Multiblock multiblock) {
         Map<String, List<List<String>>> layout = structure.getLayout();
         Map<String, Block> blocks = structure.getBlocks();
         int[] dimensions = structure.getDimensions();
 
-        // Get the block instance from the registry or wherever you have stored it
         Block controllerBlock = level.getBlockState(pos).getBlock();
         LOGGER.info("Controller Block: " + controllerBlock);
 
-        // Calculate the offset of the controller block dynamically
-        BlockPos controllerOffset = findControllerOffset(structure, controllerBlock);
+        BlockPos controllerOffset = findControllerOffset(structure, controllerBlock).orElse(BlockPos.ZERO);
         LOGGER.info("Controller Offset: " + controllerOffset);
 
-        // Adjust the starting position by the offset
         BlockPos adjustedPos = pos.offset(-controllerOffset.getX(), -controllerOffset.getY(), -controllerOffset.getZ());
         LOGGER.info("Adjusted Position: " + adjustedPos);
 
         for (int y = 0; y < dimensions[1]; y++) {
-            LOGGER.info("Checking Layer y: " + (y+1));
+            LOGGER.info("Checking Layer y: " + (y + 1));
             List<List<String>> layer = layout.get("layer" + (y + 1));
             if (layer == null) {
-                LOGGER.warn("Layer " + (y+1) + " is not defined in the layout");
+                LOGGER.warn("Layer " + (y + 1) + " is not defined in the layout");
                 return false;
             }
             for (int z = 0; z < dimensions[2]; z++) {
-                LOGGER.info("Checking Layer z: " + (z+1));
+                LOGGER.info("Checking Layer z: " + (z + 1));
                 for (int x = 0; x < dimensions[0]; x++) {
-                    LOGGER.info("Checking Layer x: " + (x+1));
-                    String blockKey = layer.get(z).get(x);
-                    if (!blockKey.equals(" ")) {
-                        Block expectedBlock = blocks.get(blockKey);
-                        if (expectedBlock == null) {
-                            LOGGER.warn("Block key " + blockKey + " is not defined in the blocks map");
-                            return false;
-                        }
-                        BlockState stateInWorld = world.getBlockState(adjustedPos.offset(x, y, z));
-                        LOGGER.info("Checking Position: " + adjustedPos.offset(x, y, z));
-                        LOGGER.info("Expected Block: " + expectedBlock);
-                        LOGGER.info("Block in World: " + stateInWorld.getBlock());
-                        if (!stateInWorld.is(expectedBlock)) {
-                            LOGGER.warn("Mismatch found at position: " + adjustedPos.offset(x, y, z) + " Expected: " + expectedBlock + " Found: " + stateInWorld.getBlock());
-                            return false;
-                        }
+                    LOGGER.info("Checking Layer x: " + (x + 1));
+                    if (!validateBlock(world, adjustedPos, x, y, z, blocks, layer)) {
+                        return false;
                     }
                 }
             }
@@ -184,8 +172,26 @@ public abstract class BaseBlockEntityMagneticraft2 extends BlockEntity implement
         LOGGER.info("Structure matched successfully.");
         return true;
     }
-
-    public BlockPos findControllerOffset(MultiblockStructure structure, Block controllerBlock) {
+    private boolean validateBlock(Level world, BlockPos adjustedPos, int x, int y, int z, Map<String, Block> blocks, List<List<String>> layer) {
+        String blockKey = layer.get(z).get(x);
+        if (!blockKey.equals(" ")) {
+            Block expectedBlock = blocks.get(blockKey);
+            if (expectedBlock == null) {
+                LOGGER.warn("Block key " + blockKey + " is not defined in the blocks map");
+                return false;
+            }
+            BlockState stateInWorld = world.getBlockState(adjustedPos.offset(x, y, z));
+            LOGGER.info("Checking Position: " + adjustedPos.offset(x, y, z));
+            LOGGER.info("Expected Block: " + expectedBlock);
+            LOGGER.info("Block in World: " + stateInWorld.getBlock());
+            if (!stateInWorld.is(expectedBlock)) {
+                LOGGER.warn("Mismatch found at position: " + adjustedPos.offset(x, y, z) + " Expected: " + expectedBlock + " Found: " + stateInWorld.getBlock());
+                return false;
+            }
+        }
+        return true;
+    }
+    public Optional<BlockPos> findControllerOffset(MultiblockStructure structure, Block controllerBlock) {
         Map<String, List<List<String>>> layout = structure.getLayout();
         int[] dimensions = structure.getDimensions();
         Map<String, Block> blocks = structure.getBlocks();
@@ -196,13 +202,12 @@ public abstract class BaseBlockEntityMagneticraft2 extends BlockEntity implement
                 for (int x = 0; x < dimensions[0]; x++) {
                     String blockKey = layer.get(z).get(x);
                     if (blocks.get(blockKey) == controllerBlock) {
-                        // Calculate the offset based on the relative position of the controller block in the structure
-                        return new BlockPos(x - dimensions[0] / 2 +1, y - dimensions[1] / 2+1, z - dimensions[2] / 2+1);
+                        return Optional.of(new BlockPos(x - dimensions[0] / 2 + 1, y - dimensions[1] / 2 + 1, z - dimensions[2] / 2 + 1));
                     }
                 }
             }
         }
-        return BlockPos.ZERO; // Default offset (0, 0, 0) if the controller block is not found in the structure
+        return Optional.empty();
     }
 
     @Nonnull
